requirements.txt
Flask
pandas
python-jobspy  # Specify the correct version if required
gunicorn
waitress